+++ src/main/java/com/sparta/myblogserver/dto/response/SuccessResponse.java
@@ -1,4 +1,4 @@
-package com.sparta.myblogserver.controller.response;
+package com.sparta.myblogserver.dto.response;
 
 import org.springframework.http.HttpStatus;
